chore: moved to platform/assets
