+++ src/platform/assets/composables/useAssetSelectionStore.ts
